refactor: move SplitFactStr test from curve to internal/utils package
