+++ internal/utils/keccak_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 // TestGetSelectorFromName tests the GetSelectorFromName function.
@@ -71,5 +72,55 @@ func TestComputeFact(t *testing.T) {
 	for _, tt := range testFacts {
 		hash := ComputeFact(tt.programHash, tt.programOutput)
 		assert.Equal(t, tt.expected, hash)
+	}
+}
+
+// TestSplitFactStr is a test function that tests the SplitFactStr function.
+//
+// It verifies the behavior of the SplitFactStr function by providing different inputs and checking the output.
+// The function takes no parameters and returns no values.
+//
+// Parameters:
+//   - t: The testing.T object for running the test
+//
+// Returns:
+//
+//	none
+func TestSplitFactStr(t *testing.T) {
+	type tescase struct {
+		input string
+		h     string
+		l     string
+		err   bool
+	}
+	data := []tescase{
+		{
+			input: "0x3",
+			h:     "0x0",
+			l:     "0x3",
+		},
+		{
+			input: "0x300000000000000000000000000000000",
+			h:     "0x3",
+			l:     "0x0",
+		},
+		{
+			input: "11111111111111111111111111111111111111111111111111111111111111010",
+			err:   true,
+		},
+		{
+			input: "X",
+			err:   true,
+		},
+	}
+	for _, d := range data {
+		l, h, err := SplitFactStr(d.input)
+		if d.err {
+			require.Error(t, err)
+		} else {
+			require.NoError(t, err)
+			assert.Equal(t, d.l, l)
+			assert.Equal(t, d.h, h)
+		}
 	}
 }
